requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
Werkzeug==1.0.1
